requirements.txt
pillow
opencv-python
numpy
cairosvg
